Fix timing corner case to avoid panic()
If session.Close() takes a while and there is code that executes after calling Close() that also takes a while, the right combination of delays will result in a panic in shutdown() when it writes to the closedChannel.
I have created a demo in new folder "close_corner_case_problem" to show the problem. Read and follow the comments before main() in demo_channel_closed.go